consolidate closinghttp code to reduce redundancy
